+++ lib/data-struct/vector.h
@@ -6,17 +6,17 @@
 #define LIBTRACE_VECTOR_H
 
 typedef struct libtrace_vector {
-	int max_size;
-	int size;
-	int element_size;
+	size_t max_size;
+	size_t size;
+	size_t element_size;
 	char *elements; // Means we can use array indexing
 	pthread_mutex_t lock;
 } libtrace_vector_t;
 
-DLLEXPORT inline void libtrace_vector_init(libtrace_vector_t *v, int element_size);
+DLLEXPORT inline void libtrace_vector_init(libtrace_vector_t *v, size_t element_size);
 DLLEXPORT inline void libtrace_vector_push_back(libtrace_vector_t *v, void *d);
-DLLEXPORT inline int libtrace_vector_get_size(libtrace_vector_t *v);
-DLLEXPORT inline int libtrace_vector_get(libtrace_vector_t *v, int location, void *d);
+DLLEXPORT inline size_t libtrace_vector_get_size(libtrace_vector_t *v);
+DLLEXPORT inline int libtrace_vector_get(libtrace_vector_t *v, size_t location, void *d);
 DLLEXPORT inline void libtrace_vector_append(libtrace_vector_t *dest, libtrace_vector_t *src);
 DLLEXPORT inline void libtrace_vector_destroy(libtrace_vector_t *v);
 DLLEXPORT inline void libtrace_zero_vector(libtrace_vector_t *v);
